Modify pinout and layouts for Nyquist
