+++ keyboards/nyquist/Makefile
@@ -1,4 +1,4 @@
-SUBPROJECT_DEFAULT = rev2
+SUBPROJECT_DEFAULT = rev1
 
 ifndef MAKEFILE_INCLUDED
 	include ../../Makefile
