--- keyboards/nyquist/lets_split.c
@@ -1,1 +1,0 @@
-#include "lets_split.h"
