+++ keyboards/nyquist/nyquist.c
@@ -1,0 +1,1 @@
+#include "nyquist.h"
