+++ keyboards/nyquist/rev1/config.h
@@ -21,23 +21,21 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "config_common.h"
 
 /* USB Device descriptor parameter */
-#define VENDOR_ID       0xFEED
-#define PRODUCT_ID      0x3060
-#define DEVICE_VER      0x0001
-#define MANUFACTURER    Wootpatoot
-#define PRODUCT         Lets Split v1
-#define DESCRIPTION     A split keyboard for the cheap makers
+#define VENDOR_ID       0xCEEB
+#define PRODUCT_ID      0x1156
+#define DEVICE_VER      0x0100
+#define MANUFACTURER    Keebio
+#define PRODUCT         The Nyquist Keyboard
+#define DESCRIPTION     Split 60 percent ortholinear keyboard
 
 /* key matrix size */
 // Rows are doubled-up
-#define MATRIX_ROWS 8
+#define MATRIX_ROWS 10
 #define MATRIX_COLS 6
 
 // wiring of each half
-#define MATRIX_ROW_PINS { B5, B4, E6, D7 }
-#define MATRIX_COL_PINS { F4, F5, F6, F7, B1, B3 }
-// #define MATRIX_COL_PINS { B3, B1, F7, F6, F5, F4 } //uncomment this line and comment line above if you need to reverse left-to-right key order
-
+#define MATRIX_ROW_PINS { D4, D7, E6, B4, B5 }
+#define MATRIX_COL_PINS { F6, F7, B1, B3, B2, B6 }
 
 #define CATERINA_BOOTLOADER
