--- keyboards/nyquist/rev2/Makefile
@@ -1,3 +1,0 @@
-ifndef MAKEFILE_INCLUDED
-	include ../../Makefile
-endif
