--- keyboards/nyquist/rev2/rules.mk
@@ -1,5 +1,0 @@
-BACKLIGHT_ENABLE = no
-
-ifndef QUANTUM_DIR
-	include ../../../Makefile
-endif
